move casn private function declarations to a new header
